% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{run_metropolis_MCMC_betas}
\alias{run_metropolis_MCMC_betas}
\title{Internal: Metropolis-Hastings MCMC backend}
\arguments{
\item{N, gamma_prior, iterations, x_vars, y, use_data_priors, user_fixed_priors, dist, epsilon, distance_metric, size_start, theta_start}{Internal.}
}
\value{
A list of MCMC results.
}
\description{
Internal C++ MCMC driver. Not for direct use.
}
\keyword{internal}
