% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data2.R
\docType{data}
\name{test_data2}
\alias{test_data2}
\title{Simulated Test Data for MCMC Examples}
\format{
A data frame with 1275 rows and 9 variables:
\describe{
  \item{start}{Numeric predictor 1.}
  \item{end.i.}{Numeric predictor 2.}
  \item{start.j.}{Numeric predictor 3.}
  \item{end}{Numeric predictor 4.}
  \item{chrom}{Character predictor 5.}
  \item{GC}{Numeric predictor 6.}
  \item{ACC}{Numeric predictor 7.}
  \item{TES}{Numeric predictor 8.}
  \item{interactions}{Numeric outcome 9.}
}
}
\source{
Generated by \code{\link{get_data}} internally.
}
\usage{
test_data2
}
\description{
A small synthetic dataframe used to illustrate how
the package works.
}
\examples{
data(test_data2)
head(test_data2)
}
\keyword{datasets}
